% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input_functions.R
\name{align_digits}
\alias{align_digits}
\title{Align the digits from the left/right of a data column and update it to the specified data frame}
\usage{
align_digits(
  indata,
  outdata,
  naming_method,
  col_name,
  align_direction = "left"
)
}
\arguments{
\item{indata}{The data column to be turned into aligned digits dataframe}

\item{outdata}{The dataframe \code{indata} will be appended (cbind) to}

\item{naming_method}{Either \code{left_aligned_column_names} or \code{right_aligned_column_names} from \code{DigitAnalysis} class}

\item{col_name}{Column name for \code{indata}}

\item{align_direction}{'left' or 'right': Create left-aligned or right-aligned digits}
}
\value{
\code{outdata}: outdata appended with indata's aligned digits
}
\description{
Align the digits from the left/right of a data column and update it to the specified data frame
}
