% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padding_test.R
\name{Benford_simulation}
\alias{Benford_simulation}
\title{Stimulates N Benford distribution datasets with matching digit places as the observed dataset. SHOULD I EXPORT THIS???!!!}
\usage{
Benford_simulation(N, freq_table, expected_mean, contingency_table)
}
\arguments{
\item{N}{The number of Benford conforming datasets to simulate.
\itemize{
  \item 2400 seconds for N=10,000; data dimension = 4000 x 5 total digits.
}}

\item{freq_table}{\code{freq_table} item returned from \code{get_expected_mean}}

\item{expected_mean}{\code{expected_mean} item returned from \code{get_expected_mean}}

\item{contingency_table}{The user-input probability table of arbitrary distribution. Overwrites \code{distribution} if not NA.
Must be a dataframe of the form as \code{benford_table}. Defaulted to NA.
\itemize{
  \item Check out \code{load(file = "data/benford_table.RData")} to see the format of \code{benford_table}
}}
}
\value{
A dataframe for the mean of each simulated dataset in every digit place
}
\description{
Stimulates N Benford distribution datasets with matching digit places as the observed dataset. SHOULD I EXPORT THIS???!!!
}
