% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_square_goodness_of_fit_functions.R
\name{chi_square_gof}
\alias{chi_square_gof}
\title{Performs chi square test for goodness of fit test. Plots \code{expected_table} with 2D and/or 3D histograms accordingly if specified.}
\usage{
chi_square_gof(
  observed_table,
  expected_table,
  freq = TRUE,
  suppress_low_N = FALSE
)
}
\arguments{
\item{observed_table}{Observation table for chi square test}

\item{expected_table}{Expected table for chi square test}

\item{freq}{TRUE or FALSE: Whether the input expected table is in decimal.}

\item{suppress_low_N}{TRUE or FALSE: If TRUE, suppress columns in expected table
if at least one cell in that column has expected value < 5. Default to FALSE.}
}
\value{
p-value for chi square goodness of fit test
}
\description{
Performs chi square test for goodness of fit test. Plots \code{expected_table} with 2D and/or 3D histograms accordingly if specified.
}
