% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digit_pair_test.R
\name{digit_pairs_test}
\alias{digit_pairs_test}
\title{Performs terminal digit pair binomial test vs uniform distribution (Benford’s Law)}
\usage{
digit_pairs_test(
  digitdata,
  data_columns = "all",
  omit_05 = NA,
  min_length = 3,
  break_out = NA,
  test_alternative = "less",
  break_out_grouping = NA,
  plot = TRUE
)
}
\arguments{
\item{digitdata}{A object of class \code{DigitAnalysis}.}

\item{data_columns}{The names of numeric columns of data to be analyzed. Default can be 'all', where using all data columns in \code{numbers} df in \code{digitdata};
an array of column names, as characters; a single column name, as character.}

\item{omit_05}{Whether to omit 0 or both 0 and 5. If omit both 0 and 5, pass in c(0,5) or c(5,0); if omit only 0 pass in 0 or c(0); if omit neither, pass in NA. Default to NA.}

\item{min_length}{The minimum length of the numbers to be analyzed in digit pair test. Must be an integer >= 2. Default to 3.}

\item{break_out}{\itemize{
  \item The data column (non-numeric!) to split up the dataset based on different categories in the column if specified as an character.
  \item The first division (usually x-axis) shown in plots.
  \item Default to NA.
}}

\item{test_alternative}{The test alternative for binomial test performed.
\itemize{
  \item 'l' or 'less': The default; test against the alternative: less than the expected frequency
  \item 't' or 'two.sided': Test against the alternative: more or less than the expected frequency
}}

\item{break_out_grouping}{A list of arrays, or defaulted to NA. Only effective if \code{break_out} is not NA.
\itemize{
  \item Each the names of the elements in the list is the break_out name
  \item Each array contains the values belonging to that break_out
  \item If it is remain as NA as default, while \code{break_out} is not NA, then \code{break_out_grouping} will default to every individual item in
  \code{break_out} will be in a separate group.
}}

\item{plot}{TRUE or FALSE or 'Save': If TRUE, display the plots and return them. If 'Save', return the plots but suppress display. If FALSE, no plot is produced. Default to TRUE.}
}
\value{
\itemize{
  \item A table of p-values for terminal digit pair test on each category
  \item A table of sample sizes for terminal digit pair test on each category
  \item Plots for each category if \code{plot = TRUE or 'Save'}
}
}
\description{
Performs terminal digit pair binomial test vs uniform distribution (Benford’s Law)
}
\examples{
digit_pairs_test(digitdata, data_columns='all')
digit_pairs_test(digitdata, data_columns=c('col_name1', 'col_name2'))
digit_pairs_test(digitdata, data_columns='all', omit_05=NA, min_length=5)
digit_pairs_test(digitdata, data_columns='all', omit_05=0, break_out='col_name')
}
