% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_square_goodness_of_fit_functions.R
\name{drop_disqualified_columns}
\alias{drop_disqualified_columns}
\title{Drops disqualified columns in both expected and observed table if in expected table at least one cell
in that column has expected value < 5, conforming to the principles of chi square test.}
\usage{
drop_disqualified_columns(observed_table, expected_table, freq = TRUE)
}
\arguments{
\item{observed_table}{Observation table for chi square test}

\item{expected_table}{Expected table for chi square test}

\item{freq}{TRUE or FALSE: Whether the input expected table is in decimal.}
}
\value{
A list with modified \code{observed_table} and \code{expected_table}
}
\description{
Drops disqualified columns in both expected and observed table if in expected table at least one cell
in that column has expected value < 5, conforming to the principles of chi square test.
}
