% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_digit_test_helper_functions.R
\name{drop_last_digit_places}
\alias{drop_last_digit_places}
\title{Remove last digit from a left-aligned/right-aligned digits table}
\usage{
drop_last_digit_places(single_column_digits, align_direction = "left")
}
\arguments{
\item{single_column_digits}{left-aligned/right-aligned digits table for a single numeric column}

\item{align_direction}{'left' or 'right': specify whether left-aligned or right-aligned digits table \code{single_column_digits} is}
}
\value{
\code{single_column_digits} without last digits
}
\description{
Remove last digit from a left-aligned/right-aligned digits table
}
