% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeat_test.R
\name{find_percent_repeats}
\alias{find_percent_repeats}
\title{Finds the percent of repeats in the given data based on given definition of a repeat (what columns need to match)}
\usage{
find_percent_repeats(data, data_column, rounding_patterns_to_omit)
}
\arguments{
\item{data}{The The dataframe to compute percent repeat on}

\item{data_column}{The data column to omit rounding numbers upon. Must be a single column.}

\item{rounding_patterns_to_omit}{The patterns to be counted as rounding digits to skip. Defaulted to NA.
\itemize{
  \item An array of characters such as c('0','00','000','5','50','500', '75', '25').
  \item \code{n_zeros_pattern} might be helpful for generating strings of 0s.
}}
}
\value{
The percentage of repeated numbers in input data
}
\description{
Finds the percent of repeats in the given data based on given definition of a repeat (what columns need to match)
}
