% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_digit_test_helper_functions.R
\name{format_p_values}
\alias{format_p_values}
\title{Display p-value as decimals when p-value >= 0.001, and as scientific when p-value < 0.001.}
\usage{
format_p_values(p_value)
}
\arguments{
\item{p_value}{The p-value computed}
}
\value{
p-value with better formatting
}
\description{
Display p-value as decimals when p-value >= 0.001, and as scientific when p-value < 0.001.
}
