% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digit_pair_test.R
\name{freq_true}
\alias{freq_true}
\title{Computes the theoratical frequency of terminal digit pairs}
\usage{
freq_true(omit_05)
}
\arguments{
\item{omit_05}{Whether to omit 0 or 5. If omit 0, 00 will be excluded. If omit 5, 05 will be excluded.}
}
\value{
The theoratical frequency of terminal digit pairs
}
\description{
Computes the theoratical frequency of terminal digit pairs
}
