% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padding_test.R
\name{get_benford_mean}
\alias{get_benford_mean}
\title{Obtains Benford mean in each digit place after specifying omitting 0 and 5 or not}
\usage{
get_benford_mean(contingency_table, omit_05 = c(0, 5))
}
\arguments{
\item{contingency_table}{The user-input probability table of arbitrary distribution. Overwrites \code{distribution} if not NA.
Must be a dataframe of the form as \code{benford_table}. Defaulted to NA.
\itemize{
  \item Check out \code{load(file = "data/benford_table.RData")} to see the format of \code{benford_table}
}}

\item{omit_05}{Whether to omit 0 or both 0 and 5. If omit both 0 and 5, pass in c(0,5) or c(5,0); if omit only 0 pass in 0 or c(0); if omit neither, pass in NA. Default to NA.}
}
\value{
A table for Benford mean in each digit place after removing 0 and/or 5 if desired
}
\description{
Obtains Benford mean in each digit place after specifying omitting 0 and 5 or not
}
