% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_square_goodness_of_fit_functions.R
\name{get_df}
\alias{get_df}
\title{Compute the degrees of freedom for the input chi square test table using formula matching that of excel.
Formula: df = (r-1)(c-1). If table has only one column, df = r - 1.}
\usage{
get_df(table)
}
\arguments{
\item{table}{Either the expected or the observation table for the chi square test}
}
\value{
Degrees of freedom
}
\description{
Compute the degrees of freedom for the input chi square test table using formula matching that of excel.
Formula: df = (r-1)(c-1). If table has only one column, df = r - 1.
}
