% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_digit_test_helper_functions.R
\name{grab_desired_aligned_columns}
\alias{grab_desired_aligned_columns}
\title{Fetches the left-aligned/right-aligned data columns for analysis, drop the first and last digit places if desired}
\usage{
grab_desired_aligned_columns(
  digitdata,
  data_columns,
  skip_first_digit = FALSE,
  skip_last_digit = FALSE,
  align_direction = "left"
)
}
\arguments{
\item{digitdata}{A object of class \code{DigitAnalysis}.}

\item{data_columns}{The names of numeric columns of data to be analyzed. Default can be 'all', where using all data columns in \code{numbers} df in \code{digitdata};
an array of column names, as characters; a single column name, as character.}

\item{skip_first_digit}{TRUE or FALSE: If TRUE, skip first digit place before analysis. Default to FALSE.}

\item{skip_last_digit}{TRUE or FALSE: If TRUE, skip last digit place before analysis, since we don't want tests to overlap. Default to FALSE.
\code{skip_last_digit} should overwrite \code{digit_places} and \code{skip_first_digits}.}
}
\value{
A list with
\itemize{
  \item \code{digits_table} The left-aligned/right-aligned digits table for \code{data_columns}
  \item \code{digitdata} The \code{DigitAnalysis} object with slot 'max' updated
}
}
\description{
Fetches the left-aligned/right-aligned data columns for analysis, drop the first and last digit places if desired
}
