% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{hist_2D_variables}
\alias{hist_2D_variables}
\title{Plot 2d histogram with multiple variables as specified by typically the \code{break_out} in digit tests.
If \code{data} is row style, rownames would be the categories, and colnames would be x values.}
\usage{
hist_2D_variables(
  data,
  data_style = "row",
  xlab = "Digits",
  ylab = "Frequency",
  title = "Multi-variable 2D Barplot",
  abline = NA
)
}
\arguments{
\item{data}{The 2D dataframe to be plotted.}

\item{data_style}{The style of input \code{data}. \code{ggplot} requires columns to be the arrays for x values, y values, etc.
\itemize{
  \item If \code{data} has a row for y values, pass in 'row'.
  \item If \code{data} has a column for y values, pass in 'col'.
  \item Defaulted to 'row'.
}}

\item{xlab}{x-axis label. Defaulted to 'digits'.}

\item{ylab}{y-axis label. Defaulted to 'frequency'.}

\item{title}{Plot title. Defaulted to '2D Histogram'.}
}
\value{
A ggplot instance.
}
\description{
Plot 2d histogram with multiple variables as specified by typically the \code{break_out} in digit tests.
If \code{data} is row style, rownames would be the categories, and colnames would be x values.
}
