% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{hist_3d}
\alias{hist_3d}
\title{Plot 3D histogram. The display follows that rows will be x-axis, and columns will be y-axis.}
\usage{
hist_3d(
  data,
  digitdata,
  xlab = "Digits",
  ylab = "Digit Places",
  zlab = "Frequency",
  title = "3D Barplot",
  theta = 55,
  phi = 16,
  plot = TRUE,
  save3Dfilename = "",
  kwargs = NA
)
}
\arguments{
\item{data}{The 2D dataframe to be plotted.}

\item{digitdata}{A object of class \code{DigitAnalysis}.}

\item{xlab}{x-axis label. Defaulted to 'digits'.}

\item{ylab}{y-axis label. Defaulted to 'frequency'.}

\item{zlab}{z-axis label. Defaulted to 'frequency'.}

\item{title}{Plot title. Defaulted to '2D Histogram'.}

\item{theta}{x-y (left-right) viewing angle. Defaulted to 55.}

\item{phi}{xy-z (up-down) viewing angle. Defaulted to 16.}

\item{plot}{TRUE or FALSE or 'Save': If TRUE, display the plots and return them. If 'Save', return the plots but suppress display. If FALSE, no plot is produced. Default to TRUE.}

\item{save3Dfilename}{If specified, will save the 3D barplot to apdf named as the input name + break out and category specification. Defaulted to ''.}

\item{kwargs}{extra parameters to pass into 3D plotting; dnt use it now, error prone! Defaulted to NA. Don't try to use it!!!!!!!!!}
}
\value{
Nothing is returned. Displays a \code{plot3D} 3d plot automatically.
}
\description{
Plot 3D histogram. The display follows that rows will be x-axis, and columns will be y-axis.
}
