% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_check_function.R
\name{input_check}
\alias{input_check}
\title{Asserts the input to user accessible functions in this package are of correct type and form.}
\usage{
input_check(
  digitdata,
  contingency_table = NA,
  data_columns = NA,
  digit_places = NA,
  skip_first_digit = NA,
  omit_05 = NA,
  break_out = NA,
  break_out_grouping = NA,
  distribution = NA,
  plot = NA,
  skip_last_digit = NA,
  unpacking_rounding_column = NA,
  min_length = NA,
  duplicate_matching_cols = NA,
  category = NA,
  category_grouping = NA,
  high = NA,
  max_length = NA,
  num_digits = NA,
  N = NA,
  standard_df = NA,
  rounding_patterns = NA,
  suppress_low_N = NA,
  suppress_first_division_plots = NA,
  suppress_second_division_plots = NA,
  simulate = NA
)
}
\arguments{
\item{digitdata}{Mandatory option. Check if it is of class \code{DigitAnalysis}}

\item{contingency_table}{Defaulted to NA. Pass in to check if it is a valid input.}

\item{data_columns}{Defaulted to NA. Pass in to check if it is a valid input.}

\item{digit_places}{Defaulted to NA. Pass in to check if it is a valid input.}

\item{skip_first_digit}{Defaulted to NA. Pass in to check if it is a valid input.}

\item{omit_05}{Defaulted to NA. Pass in to check if it is a valid input.}

\item{break_out}{Defaulted to NA. Pass in to check if it is a valid input.}

\item{distribution}{Defaulted to NA. Pass in to check if it is a valid input.}

\item{plot}{Defaulted to NA. Pass in to check if it is a valid input.}

\item{skip_last_digit}{Defaulted to NA. Pass in to check if it is a valid input.}

\item{unpacking_rounding_column}{Defaulted to NA. Pass in to check if it is a valid input.}

\item{min_length}{Defaulted to NA. Pass in to check if it is a valid input.}

\item{duplicate_matching_cols}{Defaulted to NA. Pass in to check if it is a valid input.}

\item{category}{Defaulted to NA. Pass in to check if it is a valid input.}

\item{category_grouping}{Defaulted to NA. Pass in to check if it is a valid input.}

\item{high}{Defaulted to NA. Pass in to check if it is a valid input.}

\item{max_length}{Defaulted to NA. Pass in to check if it is a valid input.}

\item{num_digits}{Defaulted to NA. Pass in to check if it is a valid input.}

\item{N}{Defaulted to NA. Pass in to check if it is a valid input.}

\item{suppress_low_N}{Defaulted to NA. Pass in to check if it is a valid input.}

\item{suppress_first_division_plots}{Defaulted to NA. Pass in to check if it is a valid input.}

\item{suppress_second_division_plots}{Defaulted to NA. Pass in to check if it is a valid input.}
}
\value{
Throw error with message if input is of incorrect type. Returns nothing.
}
\description{
Asserts the input to user accessible functions in this package are of correct type and form.
}
