% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Benford_table_functions.R
\name{load_Benford_table}
\alias{load_Benford_table}
\title{Loads Benford contingency table}
\usage{
load_Benford_table(table_fp)
}
\arguments{
\item{table_fp}{Filepath where the developers storing the table}
}
\value{
The loaded Benford table
}
\description{
Loads Benford contingency table
}
