% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input_functions.R
\name{make_cleaned_data}
\alias{make_cleaned_data}
\title{Clean up the numeric columns of dataframe}
\usage{
make_cleaned_data(raw_data, digit_columns)
}
\arguments{
\item{raw_data}{'raw' slot for \code{DigitAnalysis}}

\item{digit_columns}{The numeric columns to be cleaned}
}
\value{
Cleaned dataframe for \code{DigitAnalysis} slot 'cleaned'
}
\description{
Clean up the numeric columns of dataframe
}
