% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input_functions.R
\name{make_raw_data}
\alias{make_raw_data}
\title{Load raw dataframe from file or from passed-in dataframe}
\usage{
make_raw_data(filepath = NA, filetype = "csv", delim = ",", raw_df = NA)
}
\arguments{
\item{filepath}{Default to NA. If loading data using filepath, specify filepath as a string.}

\item{filetype}{Default to 'csv'. If loading data using filepath, specify either 'csv' or 'excel'.
'excel' option supports both 'xlsx' and 'xls' format.}

\item{delim}{Defaulted to ','. Can specify other delimeters as well.}

\item{raw_df}{Default to NA. If loading data using a dataframe. Pass in the dataframe instance.}
}
\value{
Raw dataframe for \code{DigitAnalysis} slot 'raw'
}
\description{
Load raw dataframe from file or from passed-in dataframe
}
