% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_digit_test_helper_functions.R
\name{make_sub_digitdata}
\alias{make_sub_digitdata}
\title{Create a sub-object of the \code{digitdata} object given the indexes of the rows that belong to the sub-object.}
\usage{
make_sub_digitdata(digitdata, indexes)
}
\arguments{
\item{digitdata}{A object of class \code{DigitAnalysis}.}

\item{indexes}{The indexes of the rows that belong to the sub-object.}
}
\value{
A sub-object of input \code{digitdata}.
}
\description{
Create a sub-object of the \code{digitdata} object given the indexes of the rows that belong to the sub-object.
}
