% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input_functions.R
\name{max_length}
\alias{max_length}
\title{Find the length of the largest number in a data column}
\usage{
max_length(data)
}
\arguments{
\item{data}{Data column analyzed}
}
\value{
Maximum length of the largest number in the data column
}
\description{
Find the length of the largest number in a data column
}
