% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{number_ticks}
\alias{number_ticks}
\title{StackOverflow method to specify number of ticks for \code{ggplot::scale_y_continuous()}}
\usage{
number_ticks(n)
}
\arguments{
\item{n}{Number of ticks desired}
}
\value{
Nothing is returned.
}
\description{
StackOverflow method to specify number of ticks for \code{ggplot::scale_y_continuous()}
}
