% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeat_test.R
\name{omit_rounded_numbers}
\alias{omit_rounded_numbers}
\title{Omits the numbers that are qualified as rounded numbers specified by \code{rounding_patterns_to_omit}.}
\usage{
omit_rounded_numbers(data, data_columns, rounding_patterns_to_omit)
}
\arguments{
\item{data}{The The dataframe to compute percent repeat on}

\item{data_columns}{The names of numeric columns of data to be analyzed. Default can be 'all', where using all data columns in \code{numbers} df in \code{digitdata};
an array of column names, as characters; a single column name, as character.}

\item{rounding_patterns_to_omit}{The patterns to be counted as rounding digits to skip. Defaulted to NA.
\itemize{
  \item An array of characters such as c('0','00','000','5','50','500', '75', '25').
  \item \code{n_zeros_pattern} might be helpful for generating strings of 0s.
}}
}
\value{
\code{data} without round numbers.
}
\description{
Omits the numbers that are qualified as rounded numbers specified by \code{rounding_patterns_to_omit}.
}
