% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padding_test.R
\name{padding_test}
\alias{padding_test}
\title{Performs padding test vs simulations of Benford conforming datasets via percentile}
\usage{
padding_test(
  digitdata,
  data_columns = "all",
  max_length = 8,
  num_digits = 5,
  N = 10000,
  simulate = TRUE,
  omit_05 = NA,
  break_out = NA,
  break_out_grouping = NA,
  category = NA,
  category_grouping = NA,
  distribution = "Benford",
  contingency_table = NA,
  suppress_first_division_plots = NA,
  plot = TRUE
)
}
\arguments{
\item{digitdata}{A object of class \code{DigitAnalysis}.}

\item{data_columns}{The names of numeric columns of data to be analyzed. Default can be 'all', where using all data columns in \code{numbers} df in \code{digitdata};
an array of column names, as characters; a single column name, as character.}

\item{max_length}{The length of the longest numbers considered. Defaulted to 8.}

\item{num_digits}{The total number of digits aligned from the right to be analyzed. Defaulted to 5, meaning analyzing digit place 1s to 10ks.}

\item{N}{The number of Benford conforming datasets to simulate.
\itemize{
  \item 2400 seconds for N=10,000; data dimension = 4000 x 5 totla digits.
}}

\item{simulate}{TRUE or FALSE: If TRUE, will stimulate the datasets and generate p-value. If FALSE, only produces \code{diff_in_mean} and plots.
Overwrites \code{N}.}

\item{omit_05}{Whether to omit 0 or both 0 and 5. If omit both 0 and 5, pass in c(0,5) or c(5,0); if omit only 0 pass in 0 or c(0); if omit neither, pass in NA. Default to NA.}

\item{break_out}{\itemize{
  \item The data column (non-numeric!) to split up the dataset based on different categories in the column if specified as an character.
  \item The first division (usually x-axis) shown in plots.
  \item Default to NA.
}}

\item{break_out_grouping}{A list of arrays, or defaulted to NA. Only effective if \code{break_out} is not NA.
\itemize{
  \item Each the names of the elements in the list is the break_out name
  \item Each array contains the values belonging to that break_out
  \item If it is remain as NA as default, while \code{break_out} is not NA, then \code{break_out_grouping} will default to every individual item in
  \code{break_out} will be in a separate group.
}}

\item{category}{The column for splitting the data into sectors for separate analysis. The second division (usually variables) shown in plots.}

\item{category_grouping}{A list of arrays, or defaulted to NA. Only effective if \code{category} is not NA.
\itemize{
  \item Each the names of the elements in the list is the category name
  \item Each array contains the values belonging to that category
  \item If it is remain as NA as default, while \code{category} is not NA, then \code{category_grouping} will default to every individual item in
  \code{category} will be in a separate group.
  \item e.g. \code{category_grouping = list(group_1=c(category_1, category_2, ...), group_2=c(category_10, ...), group_3=c(...))}
}}

\item{distribution}{'Benford' or 'Uniform'. Case insensitive. Specifies the distribution the chi square test is testing against. Default to 'Benford'.}

\item{contingency_table}{The user-input probability table of arbitrary distribution. Overwrites \code{distribution} if not NA.
Must be a dataframe of the form as \code{benford_table}. Defaulted to NA.
\itemize{
  \item Check out \code{load(file = "data/benford_table.RData")} to see the format of \code{benford_table}
}}

\item{suppress_first_division_plots}{TRUE or FALSE: If TRUE, suppress all plots on first and second division.
If TRUE, \code{suppress_second_division_plots} will also be set to TRUE.}

\item{plot}{TRUE or FALSE: If TRUE, skip last digit place before analysis. Default to TRUE.}
}
\value{
\itemize{
  \item A list with 5 elements
  \itemize{
    \item \code{expected_mean}: the expected mean by Benford's Law
    \item \code{observed_mean}: the mean of the input data
    \item \code{diff_in_mean}: the mean difference betweeen observed_mean and expected_mean
    \item \code{p_values}: the percentile of the observed dataset among all simulated datasets in decreasing order
    \item \code{plot}: one ggplot instance per \code{break_out}
  }
  \item Plots on \code{diff_in_mean} for each category displayed if \code{plot = TRUE}
}
}
\description{
Performs padding test vs simulations of Benford conforming datasets via percentile
}
\examples{
padding_test(digitdata, contingency_table, data_columns='all', break_out='col_name')
padding_test(digitdata, contingency_table, data_columns=c('col_name1', 'col_name2'), omit_05=NA)
padding_test(digitdata, contingency_table, data_columns='all', max_length=7, num_digits=3, omit_05=0, category='category_name', category_grouping=list(...))
padding_test(digitdata, contingency_table, data_columns='all', N=100, omit_05=NA, break_out='col_name', category='category_name', category_grouping=list(...))
}
