% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_digit_test_helper_functions.R
\name{parse_contigency_table}
\alias{parse_contigency_table}
\title{Parse the contigency table to have exclusively the desired digits and digit places}
\usage{
parse_contigency_table(
  digitdata,
  contingency_table,
  digit_places,
  skip_first_digit,
  skip_last_digit,
  omit_05
)
}
\arguments{
\item{digitdata}{A object of class \code{DigitAnalysis}.}

\item{contingency_table}{The user-input probability table of arbitrary distribution. Overwrites \code{distribution} if not NA.
Must be a dataframe of the form as \code{benford_table}. Defaulted to NA.
\itemize{
  \item Check out \code{load(file = "data/benford_table.RData")} to see the format of \code{benford_table}
}}

\item{digit_places}{The indexes of left-aligned digit places to analyze. There are three options:
\itemize{
  \item 'all': analzye all digits. Set as default.
  \item An numeric array: Perform multiple digit test on the digit places desired to analyze.
  \item A number: Perform single digit test on the digit place desired to analyze. If last digit test is desired, pass in -1 or c(-1).
}}

\item{skip_first_digit}{TRUE or FALSE: If TRUE, skip last digit place before analysis. Default to FALSE.}

\item{skip_last_digit}{TRUE or FALSE: If TRUE, skip last digit place before analysis, since we don't want tests to overlap. Default to FALSE.
\code{skip_last_digit} should overwrite \code{digit_places} and \code{skip_first_digits}.}

\item{omit_05}{Whether to omit 0 or both 0 and 5. If omit both 0 and 5, pass in c(0,5) or c(5,0); if omit only 0 pass in 0 or c(0); if omit neither, pass in NA. Default to c(0,5).}
}
\value{
Contigency table with exclusively the desired digits and digit places
}
\description{
Parse the contigency table to have exclusively the desired digits and digit places
}
