% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_digit_test_helper_functions.R
\name{parse_digit_places}
\alias{parse_digit_places}
\title{Obtain exclusively the desired digit places for left aligned digits table}
\usage{
parse_digit_places(digitdata, digits_table, digit_places)
}
\arguments{
\item{digitdata}{A object of class \code{DigitAnalysis}.}

\item{digits_table}{A left-aligned digits table}

\item{digit_places}{The indexes of left-aligned digit places to analyze. There are three options:
\itemize{
  \item 'all': analyze all digits. Set as default.
  \item An numeric array: Perform multiple digit test on the digit places desired to analyze.
  \item A number: Perform single digit test on the digit place desired to analyze. If last digit test is desired, pass in -1 or c(-1).
}}
}
\value{
Digits table with desired digit places, ready to count observation
}
\description{
Obtain exclusively the desired digit places for left aligned digits table
}
