% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_aggregate_histogram}
\alias{plot_aggregate_histogram}
\title{Plot aggregated histogram for each digit weighted average across each digit place with expected distribution.}
\usage{
plot_aggregate_histogram(
  observation_table,
  expected_table,
  freq_digit_place,
  name
)
}
\arguments{
\item{observation_table}{Observation table for chi square test in frequency.}

\item{expected_table}{Expected table for chi square test in frequency.}

\item{freq_digit_place}{The frequency of each digit place in data. N in digit place i / total N.}

\item{name}{Title for the plot.}
}
\value{
A ggplot instance for aggregated histogram.
}
\description{
Plot aggregated histogram for each digit weighted average across each digit place with expected distribution.
}
