% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_all_digit_test}
\alias{plot_all_digit_test}
\title{Plots the relevant plots for obseravtion table in \code{all_digits_test}.}
\usage{
plot_all_digit_test(
  digitdata,
  observation_table,
  expected_table,
  digit_places,
  title = ""
)
}
\arguments{
\item{digitdata}{A object of class \code{DigitAnalysis}.}

\item{observation_table}{Observation table for chi square test}

\item{digit_places}{The indexes of left-aligned digit places to analyze. There are three options:
\itemize{
  \item 'all': analzye all digits. Set as default.
  \item An numeric array: Perform multiple digit test on the digit places desired to analyze.
  \item A number: Perform single digit test on the digit place desired to analyze. If last digit test is desired, pass in -1 or c(-1).
}}

\item{title}{The title for the plot after automatically generating the name for the test: either single digit test or all digit test.}
}
\value{
Nothing is returned. Displays plots automatically.
}
\description{
Plots the relevant plots for obseravtion table in \code{all_digits_test}.
}
