% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input_functions.R
\name{process_digit_data}
\alias{process_digit_data}
\title{Create an object instance for \code{DigitAnalysis}. Parse and clean the data for digit analysis.}
\usage{
process_digit_data(
  filepath = NA,
  digit_columns = NA,
  filetype = "csv",
  delim = ",",
  raw_df = NA
)
}
\arguments{
\item{filepath}{Default to NA. If loading data using filepath, specify filepath as a string.}

\item{digit_columns}{All potential datra columns to be analyzed. Defaulted to NA. Can be specified as any of
\itemize{
  \item \code{digit_columns} = 'col_name'
  \item \code{digit_columns} = c('col_name')
  \item \code{digit_columns} = c('col_name1','col_name2', ...)
}}

\item{filetype}{Default to 'csv'. If loading data using filepath, specify either 'csv' or 'excel'.
'excel' option supports both 'xlsx' and 'xls' format.}

\item{delim}{Defaulted to ','. Can specify other delimeters as well.}

\item{raw_df}{Default to NA. If loading data using a dataframe. Pass in the dataframe instance.}
}
\value{
An object in \code{DigitAnalysis}
}
\description{
Create an object instance for \code{DigitAnalysis}. Parse and clean the data for digit analysis.
}
\examples{
process_digit_data('col_name', filepath='~/filename.csv')
process_digit_data('col_name', filepath='~/filename.xlsx', filetype='excel', delim=',')
process_digit_data('col_name', raw_df=my_dataframe)
}
