% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ensminger_Leder_Luis_10_tests.R
\name{replicate_ELL_10_tests}
\alias{replicate_ELL_10_tests}
\title{Replicates Ensingmer & Leder-Luis 10 Tests}
\usage{
replicate_ELL_10_tests(filepath = NA, raw_df = NA, plot = TRUE)
}
\arguments{
\item{filepath}{Default to NA. If loading data using filepath, specify filepath as a string.}

\item{raw_df}{Default to NA. If loading data using a dataframe. Pass in the dataframe instance.}

\item{plot}{TRUE or FALSE or 'Save': If TRUE, display the plots and return them. If 'Save', return the plots but suppress display. If FALSE, no plot is produced. Default to TRUE.}
}
\value{
The test statistics table and the plots for Ensingmer & Leder-Luis 10 Tests
}
\description{
Replicates Ensingmer & Leder-Luis 10 Tests
}
\examples{
replicate_ELL_10_tests(filepath=fp)
replicate_ELL_10_tests(raw_df=dataframe)
}
