% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rounding_test.R
\name{rounding_test}
\alias{rounding_test}
\title{Performs rounding test vs uniform distribution across categories in a specified data column}
\usage{
rounding_test(
  digitdata,
  rounding_patterns,
  break_out,
  data_columns = "all",
  break_out_grouping = NA,
  plot = TRUE
)
}
\arguments{
\item{digitdata}{A object of class \code{DigitAnalysis}.}

\item{rounding_patterns}{The patterns to be counted as rounding digits.
\itemize{
  \item An array of characters such as c('0','00','000','5','50','500', '75', '25').
  \item \code{n_zeros_pattern} might be helpful for generating strings of 0s.
}}

\item{break_out}{\itemize{
  \item The data column (non-numeric!) to split up the dataset based on different categories in the column if specified as an character.
  \item The first division (usually x-axis) shown in plots.
}}

\item{data_columns}{The names of numeric columns of data to be analyzed. Default can be 'all', where using all data columns in \code{numbers} df in \code{digitdata};
an array of column names, as characters; a single column name, as character.}

\item{break_out_grouping}{A list of arrays, or defaulted to NA. Only effective if \code{break_out} is not NA.
\itemize{
  \item Each the names of the elements in the list is the break_out name
  \item Each array contains the values belonging to that break_out
  \item If it is remain as NA as default, while \code{break_out} is not NA, then \code{break_out_grouping} will default to every individual item in
  \code{break_out} will be in a separate group.
}}

\item{plot}{TRUE or FALSE or 'Save': If TRUE, display the plots and return them. If 'Save', return the plots but suppress display. If FALSE, no plot is produced. Default to TRUE.}
}
\value{
\itemize{
  \item A table of p-values for rounding test on each category ordered by decreasing rounded percentage
  \item Plots for each category if \code{plot = TRUE or 'Save'}
}
}
\description{
Performs rounding test vs uniform distribution across categories in a specified data column
}
\examples{
rounding_test(digitdata)
rounding_test(digitdata, omit_05=0)
rounding_test(digitdata, omit_05=NA, break_out='col_name')
rounding_test(digitdata, data_columns=c('col_name1', 'col_name2'))
}
