% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sector_test.R
\name{sector_test}
\alias{sector_test}
\title{Performs sector test to analyze uneven distribution of percent repeats across sectors (supposed to be uniform).
A wrapper function for \code{repeat_test}.}
\usage{
sector_test(
  digitdata,
  break_out,
  category,
  category_instance_analyzing,
  data_column = NA,
  duplicate_matching_cols = "all",
  break_out_grouping = NA,
  category_grouping = NA,
  rounding_patterns_to_omit = NA,
  plot = TRUE,
  remove_all_category_visualize = FALSE
)
}
\arguments{
\item{digitdata}{A object of class \code{DigitAnalysis}.}

\item{break_out}{\itemize{
  \item The data column (non-numeric!) to split up the dataset based on different categories in the column if specified as an character.
  \item The first division (usually x-axis) shown in plots.
}}

\item{category}{The column for splitting the data into sectors for separate analysis. The second division (usually variables) shown in plots.}

\item{category_instance_analyzing}{The instance of the category to perform t test on.}

\item{data_column}{The data column to omit rounding numbers upon. Must be a single column.}

\item{duplicate_matching_cols}{An array of names of data columns two rows need to match exactly in order to be defined as a repeat.
Must include \code{data_column}. Default to 'all', meaning matching all columns in 'number' slot of \code{digitdata}.}

\item{break_out_grouping}{A list of arrays, or defaulted to NA. Only effective if \code{break_out} is not NA.
\itemize{
  \item Each the names of the elements in the list is the break_out name
  \item Each array contains the values belonging to that break_out
  \item If it is remain as NA as default, while \code{break_out} is not NA, then \code{break_out_grouping} will default to every individual item in
  \code{break_out} will be in a separate group.
}}

\item{category_grouping}{A list of arrays, or defaulted to NA. Only effective if \code{category} is not NA.
\itemize{
  \item Each the names of the elements in the list is the category name
  \item Each array contains the values belonging to that category
  \item If it is remain as NA as default, while \code{category} is not NA, then \code{category_grouping} will default to every individual item in
  \code{category} will be in a separate group.
}}

\item{rounding_patterns_to_omit}{The patterns to be counted as rounding digits to skip. Defaulted to NA.
\itemize{
  \item An array of characters such as c('0','00','000','5','50','500', '75', '25').
  \item \code{n_zeros_pattern} might be helpful for generating strings of 0s.
}}

\item{plot}{TRUE or FALSE or 'Save': If TRUE, display the plots and return them. If 'Save', return the plots but suppress display. If FALSE, no plot is produced. Default to TRUE.}

\item{remove_all_category_visualize}{TRUE or FALSE: If TRUE, remove visualization of 'All Category' dataset for plots}
}
\value{
\itemize{
  \item A table of p-values for repeat test by sector on each category
  \item Plots for each break out element with variables as categories specified by \code{category_grouping} if \code{plot = TRUE or 'Save'}
  \item If NaN is in returned table, it means that there are no occurances of the data of the sector in that category --> 0/0 in percentage
}
}
\description{
Performs sector test to analyze uneven distribution of percent repeats across sectors (supposed to be uniform).
A wrapper function for \code{repeat_test}.
}
\examples{
sector_test(digitdata, break_out='column', category='sector', category_instance_analyzing='X')
sector_test(digitdata, break_out='A', category='B', category_instance_analyzing='X', data_column='Y', rounding_patterns_to_omit='00')
}
