% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_low_test.R
\name{single_high_low_test}
\alias{single_high_low_test}
\title{Perform a single high low test. Helper function for \code{high_low_test}.}
\usage{
single_high_low_test(
  digitdata,
  contingency_table,
  data_columns,
  high,
  omit_05,
  skip_first_digit,
  skip_last_digit,
  category,
  category_grouping,
  test_type
)
}
\arguments{
\item{digitdata}{A object of class \code{DigitAnalysis}.}

\item{contingency_table}{The user-input probability table of arbitrary distribution. Overwrites \code{distribution} if not NA.
Must be a dataframe of the form as \code{benford_table}. Defaulted to NA.
\itemize{
  \item Check out \code{load(file = "data/benford_table.RData")} to see the format of \code{benford_table}
}}

\item{data_columns}{The names of numeric columns of data to be analyzed. Default can be 'all', where using all data columns in \code{numbers} df in \code{digitdata};
an array of column names, as characters; a single column name, as character.}

\item{high}{An numeric array of digits or a single number that will be classified as high digits. Defaulted to c(6,7,8,9).}

\item{omit_05}{Whether to omit 0 or both 0 and 5. If omit both 0 and 5, pass in c(0,5) or c(5,0); if omit only 0 pass in 0 or c(0); if omit neither, pass in NA. Default to c(0,5).}

\item{skip_first_digit}{TRUE or FALSE: If TRUE, skip last digit place before analysis. Default to FALSE.}

\item{skip_last_digit}{TRUE or FALSE: If TRUE, skip last digit place before analysis, since we don't want tests to overlap. Default to FALSE.
\code{skip_last_digit} should overwrite \code{digit_places} and \code{skip_first_digits}.}

\item{category}{The column for splitting the data into sectors for separate analysis. The second division (usually variables) shown in plots.}

\item{category_grouping}{A list of arrays, or defaulted to NA. Only effective if \code{category} is not NA.
\itemize{
  \item Each the names of the elements in the list is the category name
  \item Each array contains the values belonging to that category
  \item If it is remain as NA as default, while \code{category} is not NA, then \code{category_grouping} will default to every individual item in
  \code{category} will be in a separate group.
  \item e.g. \code{category_grouping = list(group_1=c(category_1, category_2, ...), group_2=c(category_10, ...), group_3=c(...))}
}}

\item{test_type}{Specifies whether to perform a binomial test on high vs low digit frequency weighted averaged across digit places with "binom",
or a chi square test on high vs. low by each digit place with "chisq". Defaulted to "binom".}
}
\value{
p_values table of high low test for input data from \code{digitdata}.
}
\description{
Perform a single high low test. Helper function for \code{high_low_test}.
}
