% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padding_test.R
\name{single_padding_test}
\alias{single_padding_test}
\title{Perform a single padding test. Helper function for \code{padding_test}.}
\usage{
single_padding_test(
  digitdata,
  contingency_table,
  data_columns,
  max_length,
  num_digits,
  N,
  omit_05,
  category,
  category_grouping,
  simulate
)
}
\arguments{
\item{digitdata}{A object of class \code{DigitAnalysis}.}

\item{contingency_table}{The user-input probability table of arbitrary distribution. Overwrites \code{distribution} if not NA.
Must be a dataframe of the form as \code{benford_table}. Defaulted to NA.
\itemize{
  \item Check out \code{load(file = "data/benford_table.RData")} to see the format of \code{benford_table}
}}

\item{data_columns}{The names of numeric columns of data to be analyzed. Default can be 'all', where using all data columns in \code{numbers} df in \code{digitdata};
an array of column names, as characters; a single column name, as character.}

\item{max_length}{The length of the longest numbers considered. Defaulted to 8.}

\item{num_digits}{The total number of digits aligned from the right to be analyzed. Defaulted to 5, meaning analyzing digit place 1s to 10ks.}

\item{N}{The number of Benford conforming datasets to simulate.
\itemize{
  \item 2400 seconds for N=10,000; data dimension = 4000 x 5 totla digits.
}}

\item{omit_05}{Whether to omit 0 or both 0 and 5. If omit both 0 and 5, pass in c(0,5) or c(5,0); if omit only 0 pass in 0 or c(0); if omit neither, pass in NA. Default to NA.}

\item{category}{The column for splitting the data into sectors for separate analysis. The second division (usually variables) shown in plots.}

\item{category_grouping}{A list of arrays, or defaulted to NA. Only effective if \code{category} is not NA.
\itemize{
  \item Each the names of the elements in the list is the category name
  \item Each array contains the values belonging to that category
  \item If it is remain as NA as default, while \code{category} is not NA, then \code{category_grouping} will default to every individual item in
  \code{category} will be in a separate group.
  \item e.g. \code{category_grouping = list(group_1=c(category_1, category_2, ...), group_2=c(category_10, ...), group_3=c(...))}
}}

\item{simulate}{TRUE or FALSE: If TRUE, will stimulate the datasets and generate p-value. If FALSE, only produces \code{diff_in_mean} and plots.
Overwrites \code{N}.}
}
\value{
A list of padding test results for input data from \code{digitdata}.
}
\description{
Perform a single padding test. Helper function for \code{padding_test}.
}
