% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_table.R
\docType{data}
\name{uniform_table}
\alias{uniform_table}
\title{The dataset for contingency table of each digit in each digit place conforming to uniform distribution.}
\format{
An object of class \code{data.frame} with 10 rows and 100 columns.
}
\usage{
uniform_table
}
\description{
A data frame with 10 rows and 100 columns. Entry [i, j] corresponds to the freqency of digit i in digit place j in decimals.
}
\keyword{datasets}
