% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_all_digit_test}
\alias{plot_all_digit_test}
\title{Plots the relevant plots for obseravtion table in \code{all_digits_test}.}
\usage{
plot_all_digit_test(
  digitdata,
  observation_table,
  expected_table,
  digit_places,
  title = "",
  plot = TRUE,
  save3Dfilename = "",
  kwargs = NA
)
}
\arguments{
\item{digitdata}{A object of class \code{DigitAnalysis}.}

\item{observation_table}{Observation table for chi square test}

\item{expected_table}{Expected table for chi square test}

\item{digit_places}{The indexes of left-aligned digit places to analyze. There are three options:
\itemize{
  \item 'all': analyze all digits. Set as default.
  \item An numeric array: Perform multiple digit test on the digit places desired to analyze.
  \item A number: Perform single digit test on the digit place desired to analyze. If last digit test is desired, pass in -1 or c(-1).
}}

\item{title}{The title for the plot after automatically generating the name for the test: either single digit test or all digit test.}

\item{plot}{TRUE or FALSE or 'Save': If TRUE, display the plots and return them. If 'Save', return the plots but suppress display. If FALSE, no plot is produced. Default to TRUE.}

\item{save3Dfilename}{If specified, will save the 3D barplot to apdf named as the input name + break out and category specification. Defaulted to ''.}

\item{kwargs}{extra parameters to pass into 3D plotting; dnt use it now, error prone! Defaulted to NA. Don't try to use it!!!!!!!!!}
}
\value{
Nothing is returned. Displays plots automatically.
}
\description{
Plots the relevant plots for obseravtion table in \code{all_digits_test}.
}
