% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ensminger_Leder_Luis_10_tests.R
\name{replicate_ELL_10_tests}
\alias{replicate_ELL_10_tests}
\title{Replicates Ensingmer & Leder-Luis 10 Tests}
\usage{
replicate_ELL_10_tests(fp = NA, raw_df = NA, plot = TRUE)
}
\arguments{
\item{fp}{filepath for input data}

\item{raw_df}{dataframe for input data}

\item{plot}{TRUE or FALSE: If generate plot or not. Default to TRUE, might have bugs if set to FALSE lmao.}
}
\value{
The test statistics table and the plots for Ensingmer & Leder-Luis 10 Tests
}
\description{
Replicates Ensingmer & Leder-Luis 10 Tests
}
\examples{
replicate_ELL_10_tests(fp=filepath)
replicate_ELL_10_tests(raw_df=dataframe)
}
